refactor: employee package를 제거하고, 내부에 존재하던 classe들을 다른 package로 이동
- employee package 제작 이유는, 레이싱 게임 내에서 게임 진행과 관련된 인격화 된 객체들을 모으기 위함
- 하지만 이러한 사실은 코드 제작자만 알 수 있는 모호한 네이밍이라고 판단
- game이라는 package를 제작하고, recoder는 record, 그 외는 game package로 이동
