+++ src/main/kotlin/racinggame/domain/game/RacingGame.kt
@@ -1,6 +1,4 @@
-package racinggame.domain
-
-import racinggame.domain.employee.GameGuide
+package racinggame.domain.game
 
 interface RacingGame {
