+++ src/main/kotlin/racinggame/domain/game/RacingGameResult.kt
@@ -1,4 +1,4 @@
-package racinggame.domain
+package racinggame.domain.game
 
 import racinggame.domain.record.RacingRecordBook
