+++ src/main/kotlin/racinggame/domain/game/RealRacingGame.kt
@@ -1,10 +1,8 @@
-package racinggame.domain
+package racinggame.domain.game
 
 import racinggame.domain.car.RacingCar
 import racinggame.domain.car.factory.RacingCarFactory
-import racinggame.domain.employee.GameGuide
-import racinggame.domain.employee.GameRule
-import racinggame.domain.employee.RacingHistoryRecoder
+import racinggame.domain.record.RacingHistoryRecoder
 import racinggame.domain.field.RacingField
 import racinggame.domain.player.User
 import racinggame.domain.player.toRacer
