+++ src/main/kotlin/racinggame/view/RacingGameController.kt
@@ -1,9 +1,9 @@
 package racinggame.view
 
-import racinggame.domain.RacingGame
-import racinggame.domain.RacingGameResult
-import racinggame.domain.employee.GameGuide
-import racinggame.domain.employee.GameRule
+import racinggame.domain.game.RacingGame
+import racinggame.domain.game.RacingGameResult
+import racinggame.domain.game.GameGuide
+import racinggame.domain.game.GameRule
 import racinggame.domain.player.User
 
 class RacingGameController(
